@@ -27,17 +27,6 @@ shared_examples_for "check that basic operations with postgres works correctly f
 
   end # when try to create one record using new/save
 
-  context "when try to create many records" do
-
-    it "records created" do
-      expect { subject.create_many([
-                                     { :name => 'Alex', :company_id => 2, :created_at => DATE_NOW + 1 },
-                                     { :name => 'Aaron', :company_id => 3, :created_at => DATE_NOW + 1 }])
-      }.not_to raise_error
-    end
-
-  end # when try to create many records
-
   context "when try to find a record with the search term is id" do
 
     it "returns employee name" do
@@ -89,36 +78,6 @@ shared_examples_for "check that basic operations with postgres works correctly f
 
   end # when try to update a record with id = 1
 
-  context "when try to update a record with update_many functions" do
-
-    it "returns updated employee name" do
-      subject.update_many( {
-        { :id => 1 } => {
-            :name => 'Alex',
-            :company_id => 3,
-            :created_at => DATE_NOW
-          }
-      } )
-      expect(subject.find(1).name).to eq("Alex")
-    end
-
-    it "returns updated employee name" do
-      rows = [{
-         :id => 1,
-         :name => 'Pit',
-         :created_at => DATE_NOW
-      }]
-
-      options = {
-        :set_array => '"name = datatable.name"',
-        :where => '"#{table_name}.id = datatable.id"'
-      }
-      subject.update_many(rows, options)
-      expect(subject.find(1).name).to eq("Pit")
-    end
-
-  end # when try to update a record with update_many functions
-
   context "when try to delete a record with id = 1" do
 
     it "returns empty array" do
@@ -131,7 +90,8 @@ shared_examples_for "check that basic operations with postgres works correctly f
   context "when try to create new record outside the range of partitions" do
 
     it "raises ActiveRecord::StatementInvalid" do
-      expect { subject.create_many([{ :created_at => DATE_NOW - 1.year, :company_id => 1 }])
+      expect {
+        subject.create(created_at: DATE_NOW - 1.year, company_id: 1)
       }.to raise_error(ActiveRecord::StatementInvalid)
     end
 
