@@ -164,9 +164,8 @@ module Partitioned
       def create_partition_table(*partition_key_values)
         create_table(configurator.table_name(*partition_key_values), {
                        :id => false,
-                       :options => "INHERITS (#{configurator.parent_table_name(*partition_key_values)})"
-                     }) do |t|
-        end
+                       :options => "() INHERITS (#{configurator.parent_table_name(*partition_key_values)})"
+                     })
         add_check_constraint(*partition_key_values)
       end
 
